CHG: Moved network components in a specific folder
